requirements.txt
numpy
pandas
supabase
torch
torchvision
fastapi
uvicorn
scikit-learn
facenet-pytorch
python-dotenv
python-multipart